requirements.txt
great-expectations
pandas
pydantic
requests
python-dotenv